backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
python-dotenv==1.0.1
transformers==4.41.2
torch==2.3.1
sentencepiece==0.2.0
pymupdf==1.24.5
python-docx==1.1.2
beautifulsoup4==4.12.3
requests==2.32.3
python-multipart==0.0.9
nltk==3.8.1
numpy==1.26.4
